Modularized into separate files
Add ToString implementation on the components
